Move solid-js related code into one folder
