+++ src/settings/solid-js/settings-store.ts
@@ -1,0 +1,3 @@
+// export function createSettingsStore() {
+// return new SettingsStore()
+// }
